requirements.txt
click==8.1.7
colorama==0.4.6
joblib==1.4.0
nltk==3.8.1
numpy==1.26.4
pandas==2.2.2
prettytable==3.10.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
six==1.16.0
tqdm==4.66.2
tzdata==2024.1
wcwidth==0.2.13